pyproject.toml
[project]
name = "portia-aml-agent"
version = "0.1.0"
description = "AML agent to flag transactions and generate SAR."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.6",
    "mcp[cli]>=1.13.1",
    "portia-sdk-python[tools-browser-browserbase]>=0.6.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
